add confirmation for delete

+++ static/js/admin.js
@@ -84,14 +84,18 @@ var Admin = {
 
       }
       $(".deleteroute").click(function(){
-        $.ajax({
-          url: '/routes/' + $(this).attr("routeId"),
-          type: 'DELETE',
-          success: function(result) {
-            //Admin.populateRoutesPanel(data);
-            $.get( "/routes", Admin.populateRoutesPanel);
-          }
-        });
+        if ($(this).html() == "sure?"){
+          $.ajax({
+            url: '/routes/' + $(this).attr("routeId"),
+            type: 'DELETE',
+            success: function(result) {
+              //Admin.populateRoutesPanel(data);
+              $.get( "/routes", Admin.populateRoutesPanel);
+            }
+          });
+        }else{
+          $(this).html("sure?")
+        }
       });
 
     }
@@ -127,7 +131,7 @@ var Admin = {
     }));
     L.tileLayer('http://tile.stamen.com/toner-lite/{z}/{x}/{y}{r}.png', {
       attribution: 'Map tiles by <a href="http://stamen.com">Stamen Design</a>, under <a href="http://creativecommons.org/licenses/by/3.0">CC BY 3.0</a>. Data by <a href="http://openstreetmap.org">OpenStreetMap</a>, under <a href="http://www.openstreetmap.org/copyright">ODbL</a>.',
-      minZoom: 13 
+      minZoom: 13
     }).addTo(Admin.RoutesMap);
 
     Admin.RoutesMap.addLayer(Admin.drawnRoute);
